pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
django-tailwind = "^3.8.0"
django-browser-reload = "^1.12.1"
django-storages = "^1.14.2"
boto3 = "^1.34.82"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"
pytest-env = "^1.1.3"
django-stubs-ext = "^4.2.7"

[tool.poetry.group.dev.dependencies]
django-stubs = "^4.2.7"
pytest = "^8.1.1"
ruff = "^0.3.5"
pytest-django = "^4.8.0"
mypy = "^1.9.0"
factory-boy = "^3.3.0"
pytest-coverage = "^0.0"
ics = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
strict = true

[[tool.mypy.overrides]]
module = [
    "factory",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "backend.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
python_files = ["test_*.py"]
env = [
  "D:SECRET_KEY=django-top-secret",
  "D:ENV=test"
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["I", "E", "F", "W", "C901", "DTZ", "DJ"]

[tool.coverage.run]
omit = ['backend.asgi.py', 'backend.wsgi.py']
